+++ src/main/kotlin/com/nbadal/ktlint/KtlintAnnotator.kt
@@ -31,6 +31,7 @@ class KtlintAnnotator : ExternalAnnotator<PsiFile, List<LintError>>() {
 
             holder.createAnnotation(severity, errorRange, message).apply {
                 registerFix(KtlintFormatAction())
+                registerFix(KtlintConfigAction())
             }
         }
     }